+++ classes/event/template_cohort_added.php
@@ -1,0 +1,77 @@
+<?php
+// This file is part of Moodle - http://moodle.org/
+//
+// Moodle is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Moodle is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
+
+/**
+ * Event: a cohort was attached to a learning plan template.
+ *
+ * @package    local_dimensions
+ * @copyright  2026 Anderson Blaine
+ * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+
+namespace local_dimensions\event;
+
+use core\event\base;
+
+/**
+ * Event: a cohort was attached to a learning plan template.
+ *
+ * Core fires no event for the template-cohort relation, so the hub logs the
+ * decision itself. The 'other' payload carries templateid, cohortid and
+ * syncqueued (whether background plan generation was queued with the attach).
+ *
+ * @package    local_dimensions
+ * @copyright  2026 Anderson Blaine
+ * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+class template_cohort_added extends base {
+    /**
+     * Initialise the event static data.
+     */
+    protected function init() {
+        $this->data['objecttable'] = \core_competency\template_cohort::TABLE;
+        $this->data['crud'] = 'c';
+        $this->data['edulevel'] = self::LEVEL_OTHER;
+    }
+
+    /**
+     * Localised event name.
+     *
+     * @return string
+     */
+    public static function get_name() {
+        return get_string('eventtemplatecohortadded', 'local_dimensions');
+    }
+
+    /**
+     * Non-localised description of what happened.
+     *
+     * @return string
+     */
+    public function get_description() {
+        return "The user with id '$this->userid' attached the cohort with id '{$this->other['cohortid']}' to the "
+            . "learning plan template with id '{$this->other['templateid']}'.";
+    }
+
+    /**
+     * No restore mapping: system-scoped, never part of a course backup.
+     *
+     * @return string
+     */
+    public static function get_objectid_mapping() {
+        return base::NOT_MAPPED;
+    }
+}